# docs/openapi/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "FinalAcceptanceResponse": {
        "additionalProperties": false,
        "properties": {
          "blockers": {
            "items": {
              "type": "string"
            },
            "title": "Blockers",
            "type": "array"
          },
          "generator_revision": {
            "title": "Generator Revision",
            "type": "string"
          },
          "real_acceptance": {
            "title": "Real Acceptance",
            "type": "boolean"
          },
          "simulated": {
            "title": "Simulated",
            "type": "boolean"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "synthetic_readiness_state": {
            "title": "Synthetic Readiness State",
            "type": "string"
          },
          "window_id": {
            "title": "Window Id",
            "type": "string"
          }
        },
        "required": [
          "window_id",
          "status",
          "blockers",
          "synthetic_readiness_state",
          "simulated",
          "real_acceptance",
          "generator_revision"
        ],
        "title": "FinalAcceptanceResponse",
        "type": "object"
      },
...
      "GovernanceSignoffRequest": {
        "additionalProperties": false,
        "properties": {
          "comment": {
            "default": "",
            "maxLength": 2000,
            "title": "Comment",
            "type": "string"
          },
          "decision": {
            "enum": [
              "APPROVE",
              "VETO"
            ],
            "title": "Decision",
            "type": "string"
          },
          "role": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "role",
          "decision"
        ],
        "title": "GovernanceSignoffRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IncidentCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "severity": {
            "enum": [
              "P0",
              "P1",
              "P2",
              "P3"
            ],
            "title": "Severity",
            "type": "string"
          },
          "title": {
            "maxLength": 300,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "severity",
          "title"
        ],
        "title": "IncidentCreateRequest",
        "type": "object"
      },
      "IncidentResponse": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "incident_id": {
            "title": "Incident Id",
            "type": "string"
          },
          "row_version": {
            "default": 1,
            "title": "Row Version",
            "type": "integer"
          },
          "severity": {
            "title": "Severity",
            "type": "string"
          },
          "simulated": {
            "title": "Simulated",
            "type": "boolean"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "window_id": {
            "title": "Window Id",
            "type": "string"
          }
        },
        "required": [
          "incident_id",
          "window_id",
          "severity",
          "title",
          "state",
          "simulated"
        ],
        "title": "IncidentResponse",
        "type": "object"
      },
...
      "ObservationCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "ObservationCreateRequest",
        "type": "object"
      },
...
      "ObservationResponse": {
        "additionalProperties": false,
        "properties": {
          "ends_at": {
            "title": "Ends At",
            "type": "number"
          },
          "metrics": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Metrics",
            "type": "object"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "real_acceptance": {
            "title": "Real Acceptance",
            "type": "boolean"
          },
          "row_version": {
            "title": "Row Version",
            "type": "integer"
          },
          "simulated": {
            "title": "Simulated",
            "type": "boolean"
          },
          "starts_at": {
            "title": "Starts At",
            "type": "number"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "window_id": {
            "title": "Window Id",
            "type": "string"
          }
        },
        "required": [
          "window_id",
          "name",
          "starts_at",
          "ends_at",
          "state",
          "metrics",
          "simulated",
          "real_acceptance",
          "row_version"
        ],
        "title": "ObservationResponse",
        "type": "object"
      },
...
      "ReadinessResponse": {
        "additionalProperties": false,
        "properties": {
          "blockers": {
            "items": {
              "type": "string"
            },
            "title": "Blockers",
            "type": "array"
          },
          "real_acceptance": {
            "title": "Real Acceptance",
            "type": "boolean"
          },
          "scope_id": {
            "title": "Scope Id",
            "type": "string"
          },
          "simulated": {
            "title": "Simulated",
            "type": "boolean"
          },
          "state": {
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "scope_id",
          "state",
          "blockers",
          "simulated",
          "real_acceptance"
        ],
        "title": "ReadinessResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Enterprise RAG Knowledge Base",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/governance/incidents/{incident_id}:resolve": {
      "put": {
        "operationId": "resolve_incident_api_v1_governance_incidents__incident_id__resolve_put",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "title": "Incident Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": true,
            "schema": {
              "title": "If-Match",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Resolve Incident",
        "tags": [
          "acceptance"
        ]
      }
    },
    "/api/v1/governance/observations": {
      "post": {
        "operationId": "create_observation_api_v1_governance_observations_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ObservationCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObservationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Create Observation",
        "tags": [
          "acceptance"
        ]
      }
    },
    "/api/v1/governance/observations/{window_id}": {
      "get": {
        "operationId": "get_observation_api_v1_governance_observations__window_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "window_id",
            "required": true,
            "schema": {
              "title": "Window Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObservationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Get Observation",
        "tags": [
          "acceptance"
        ]
      }
    },
    "/api/v1/governance/observations/{window_id}/final-acceptance-report": {
      "get": {
        "operationId": "final_acceptance_report_api_v1_governance_observations__window_id__final_acceptance_report_get",
        "parameters": [
          {
            "in": "path",
            "name": "window_id",
            "required": true,
            "schema": {
              "title": "Window Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FinalAcceptanceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Final Acceptance Report",
        "tags": [
          "acceptance"
        ]
      }
    },
    "/api/v1/governance/observations/{window_id}/incidents": {
      "post": {
        "operationId": "create_incident_api_v1_governance_observations__window_id__incidents_post",
        "parameters": [
          {
            "in": "path",
            "name": "window_id",
            "required": true,
            "schema": {
              "title": "Window Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IncidentCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Create Incident",
        "tags": [
          "acceptance"
        ]
      }
    },
...
    "/api/v1/governance/observations/{window_id}/signoffs": {
      "post": {
        "operationId": "observation_signoff_api_v1_governance_observations__window_id__signoffs_post",
        "parameters": [
          {
            "in": "path",
            "name": "window_id",
            "required": true,
            "schema": {
              "title": "Window Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": true,
            "schema": {
              "title": "If-Match",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GovernanceSignoffRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Observation Signoff Api V1 Governance Observations  Window Id  Signoffs Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Observation Signoff",
        "tags": [
          "acceptance"
        ]
      }
    },
    "/api/v1/governance/observations/{window_id}:close": {
      "post": {
        "operationId": "close_observation_api_v1_governance_observations__window_id__close_post",
        "parameters": [
          {
            "in": "path",
            "name": "window_id",
            "required": true,
            "schema": {
              "title": "Window Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": true,
            "schema": {
              "title": "If-Match",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObservationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Close Observation",
        "tags": [
          "acceptance"
        ]
      }
    },
    "/api/v1/governance/observations/{window_id}:evaluate": {
      "post": {
        "operationId": "evaluate_observation_api_v1_governance_observations__window_id__evaluate_post",
        "parameters": [
          {
            "in": "path",
            "name": "window_id",
            "required": true,
            "schema": {
              "title": "Window Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": true,
            "schema": {
              "title": "If-Match",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadinessResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Evaluate Observation",
        "tags": [
          "acceptance"
        ]
      }
    },
...
  }
}